% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFullUcdpGedData.R
\name{getFullUcdpGedData}
\alias{getFullUcdpGedData}
\title{getFullUcdpGedData: Function to download full Ucdp GED data up to a user-
specified release date.}
\usage{
getFullUcdpGedData(date = Sys.Date())
}
\arguments{
\item{date}{A date object defining the reference date for retrieving the latest dataset
versions. Defaults to the current system date.}
}
\value{
A data.frame listing the latest available versions of the UCDP GED dataset.
}
\description{
Note: Final date of data coverage is always one month prior to the UCDP
release date.
}
\examples{
# Get the latest UCDP GED data using the current system date
getFullUcdpGedData()

# Get the latest UCDP GED data using a specific date
getFullUcdpGedData(as.Date("2022-12-31"))
}
