% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getUcdpMetaData.R
\name{getUcdpMetaData}
\alias{getUcdpMetaData}
\title{Get metadata from UCDP Dataset downloaded through \pkg{ucdp.api}}
\usage{
getUcdpMetaData(ucdp.df, as.text = FALSE)
}
\arguments{
\item{ucdp.df}{A data.frame containing UCDP data with expected metadata columns.}

\item{as.text}{boolean: Should result be returned as text (TRUE) or printed
to the console (FALSE)? Default: FALSE}
}
\value{
A text summary of the metadata or message printed to console.
}
\description{
This function extracts metadata from a UCDP dataset downloaded through the
API using the \pkg{ucdp.api} R package and prints the results as a text string
or console message.
}
\examples{
ucdp.df <- read.csv("your_ucdp_data_file.csv")
getUcdpMetaData(ucdp.df, as.text = TRUE)
}
