% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pingUrl.R
\name{pingUrl}
\alias{pingUrl}
\title{pingUrl: Check if URL is valid and can be accessed}
\usage{
pingUrl(url)
}
\arguments{
\item{url}{character. URL of website to be accessed}
}
\value{
error message if URL cannot be accessed
}
\description{
pingUrl: Check if URL is valid and can be accessed
}
\examples{
pingUrl("https://www.google.com/") ## Works, no error
pingUrl("https://www.goooogle.com/") ## Will not work, results in error

}
