% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkUcdpAvailable.R
\name{checkUcdpAvailable}
\alias{checkUcdpAvailable}
\title{checkUcdpAvailable: Check which UCDP datasets are currently available}
\usage{
checkUcdpAvailable(dataset, version, as.vector = FALSE)
}
\arguments{
\item{dataset}{character. Name of UCDP dataset. Needs to be one of the following:
"ucdpprioconflict", "battledeaths", "dyadic", "nonstate", "onesided" (yearly data),
or "gedevents" (daily data, released in monthly batches)
The API does not require an account/key but there are some usage limitations
For more information, see: https://ucdp.uu.se/apidocs/}

\item{version}{character. Dataset version number.
For yearly UCDP data: Format: YY.1 (e.g. 22.1 for data released in 2022)
For monthly UCDP GED data use format: YY.0.MM (e.g. 22.0.11 for data on November 2022)
For quarterly UCDP releases data use format: YY.01.YY.MM (e.g. 22.01.22.09 for data from Jan to Sep 2022)}

\item{as.vector}{boolean: Return output as multi-column data.frame (FALSE) or as simple vector (TRUE)? Default: FALSE}
}
\value{
data.frame with requested UCDP dataset
}
\description{
checkUcdpAvailable: Check which UCDP datasets are currently available
}
\examples{
## Check which monthly UCDP GED candidate datasets are available
## m.ids <- 1:12
gedc.m.check.df <- do.call(rbind, lapply(m.ids, function(x){
  checkUcdpAvailable("gedevents", sprintf("22.0.\%s", x))
  }))

## Check which quarterly UCDP GED candidate datasets are available
q.ids <- sprintf("\%02d", seq(3, 12, 3))
gedc.q.check.df <- do.call(rbind, lapply(q.ids, function(x){
  checkUcdpAvailable("gedevents", sprintf("22.01.22.\%s", x))
  }))

}
