% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFullUcdpGedData.R
\name{getFullUcdpGedData}
\alias{getFullUcdpGedData}
\title{getFullUcdpGedData: Wrapper function to download full UCDP GED data up to the
latest release (default) or a user-specified release date.}
\usage{
getFullUcdpGedData(
  date = Sys.Date(),
  candidate.only = FALSE,
  add.metadata = TRUE
)
}
\arguments{
\item{date}{A date object defining the reference date for retrieving the latest dataset
versions. Defaults to the current system date-}

\item{candidate.only}{boolean. If set to TRUE, the call will only download the latest
monthly and quarterly candidate datasets, and skip downloading the latest yearly
release of final GED data, which can take a long time to complete. Default: FALSE}

\item{add.metadata}{boolean: Add metadata variables to output dataset (TRUE)
or not (FALSE)? Includes dataset name, version number, download date. Default: TRUE}
}
\value{
A data.frame containing the latest full UCDP GED data (final & candidate
or candidate only)
}
\description{
getFullUcdpGedData: Wrapper function to download full UCDP GED data up to the
latest release (default) or a user-specified release date.
}
\examples{
# Get the latest UCDP GED data using the current system date
getFullUcdpGedData()

# Get the latest UCDP GED data using a specific date
getFullUcdpGedData(as.Date("2022-12-31"))
}
