% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getUcdpData.R
\name{getUcdpData}
\alias{getUcdpData}
\title{Download data from the Uppsala Conflict Database Program API.}
\usage{
getUcdpData(
  dataset,
  version,
  pagesize = 1000,
  max.retries = 10,
  add.metadata = TRUE
)
}
\arguments{
\item{dataset}{character. Name of UCDP dataset. Needs to be one of the following:
"ucdpprioconflict", "battledeaths", "dyadic", "nonstate", "onesided" (yearly data),
or "gedevents" (daily data, released in monthly batches)
The API does not require an account/key but there are some usage limitations
For more information, see: https://ucdp.uu.se/apidocs/}

\item{version}{character. Dataset version number.
For yearly UCDP data: Format: YY.1 (e.g. 22.1 for data released in 2022)
For monthly UCDP GED data use format: YY.0.MM (e.g. 22.0.11 for data on November 2022)
For quarterly UCDP releases data use format: YY.01.YY.MM (e.g. 22.01.22.09 for data from Jan to Sep 2022)}

\item{pagesize}{numeric. Page size of each individual query. Default: 1000. Max: 1000
The UCDP API divides dataset into N pages of size S,
the query loops over all pages until full dataset is retrieved.}

\item{max.retries}{numeric (integer). Maximum number of retry attempts for API
calls in case of failure, default: 10. Useful for managing temporary network or server issues.
Exceeding max.retries stops the function with an error.}

\item{add.metadata}{boolean: Add metadata to output dataset (TRUE) or not (FALSE).
Default: TRUE}
}
\value{
data.frame with requested UCDP dataset
}
\description{
Download data from the Uppsala Conflict Database Program API.
}
\examples{
getUcdpData(dataset = ucdpprioconflict, version = "22.1", pagesize = 100)
}
